Download, Clean and Generate Graphs From USQ ThetaProbe Data
Source: R/get_soil_moisture.R

This package automates downloading and cleaning of soil moisture data from ThetaProbs.

This is a slow process to retrieve all the files from the server for the first run. After the first run, it will be much faster to only retrieve new files.

get_soil_moisture(userpwd = NULL, path = NULL, local_dirs = NULL)

Arguments

userpwd
The login and password provided by NCEA to login via file transfer protocol (FTP)

path
File path to directory for saving a comma separated file (CSV) output. Defaults to current working directory.

local_dirs
File path to directory, which holds previous data logger data monthly file folders, which contain hourly CSV files from theta probe loggers.

Details

This function will download CSV files from the server that are not currently on the local machine in the user specified location and collate them into one CSV file for further use.

Examples

# NOT RUN {
get_soil_moisture(userpwd = "userid:password", path = NULL,
local_dirs = NULL)
# }